+++ packages/shared/src/profiler/index.ts
@@ -1,3 +1,4 @@
 export * from "./contracts";
+export * from "./experience";
 export * from "./question-bank";
 export * from "./types";